+++ src/test/java/com/jstarcraft/ai/math/algorithm/probability/UniformProbabilityTestCase.java
@@ -10,23 +10,23 @@ import org.nd4j.linalg.api.rng.distribution.Distribution;
 
 public class UniformProbabilityTestCase extends ProbabilityTestCase {
 
-	@Override
-	protected Distribution getOldFunction(long seed) {
-		Random random = new DefaultRandom(seed);
-		Distribution distribution = new org.nd4j.linalg.api.rng.distribution.impl.UniformDistribution(random, 0.4D, 4D);
-		return distribution;
-	}
+    @Override
+    protected Distribution getOldFunction(int seed) {
+        Random random = new DefaultRandom(seed);
+        Distribution distribution = new org.nd4j.linalg.api.rng.distribution.impl.UniformDistribution(random, 0.4D, 4D);
+        return distribution;
+    }
 
-	@Override
-	protected Probability getNewFunction(long seed) {
-		return new QuantityProbability(MersenneTwister.class, seed, UniformRealDistribution.class, 0.4D, 4D);
-	}
+    @Override
+    protected Probability getNewFunction(int seed) {
+        return new QuantityProbability(MersenneTwister.class, seed, UniformRealDistribution.class, 0.4D, 4D);
+    }
 
-	@Override
-	protected void assertSample(Probability newFuction, Distribution oldFunction) {
-		Number newSample = newFuction.sample();
-		Number oldSample = oldFunction.sample();
-		Assert.assertThat(newSample, CoreMatchers.equalTo(oldSample));
-	}
+    @Override
+    protected void assertSample(Probability newFuction, Distribution oldFunction) {
+        Number newSample = newFuction.sample();
+        Number oldSample = oldFunction.sample();
+        Assert.assertThat(newSample, CoreMatchers.equalTo(oldSample));
+    }
 
 }
